fix: Resolve 42 regressions from bug hunting season 1 (season 2)
Fixes regressions and newly discovered bugs introduced by the first
round of 101 bug fixes. Key changes across 42 files (+586/-194 lines):

Auth/DB: SHA-1 HMAC with CSPRNG salt (FGuid), DB schema column added
Combat: Absorb damage flow unblocked, dual healing path removed,
  validation fallback fixed, MagnumBreak/CachedActorInfo guards
Items: IsEquipment ID-range check, trade full-field snapshot comparison,
  refinement UniqueID cleanup, loot broadcast, weight projection
Monsters/NPC: SpawnLocation preserved, AI leash checks own distance,
  NPC identity-based player lookup, dialogue per-player migration,
  quest ordering for all objective types
World/Network: Zone memory leak cleanup, ClientTravel destination
  parsing in GameMode, AlwaysRelevantNode null guards, party iteration
  safety, WarpPortal cross-map travel
Character/Core: SitDown/StandUp Server RPCs, whisper parser fix,
  respawn timer stable key, delegate double-bind prevention,
  bolt SP costs, gameplay tags registration, DropTable dice fix

+++ RagnarokUE/Source/RagnarokUE/Data/ROConstants.h
@@ -70,7 +70,7 @@ namespace ROConstants
 	// ---- Currency ----
 
 	/** Maximum zeny a character can hold. */
-	constexpr int64 MaxZeny = 1000000000LL;
+	constexpr int32 MaxZeny = 1000000000;
 
 	// ---- Timing ----
